backend/microservice-scanner/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.23.2
python-nmap==0.7.1
pydantic==2.6.0
speedtest-cli==2.1.3
netifaces==0.11.0   
